refactor(core): consolidate target model package
- move shared target-model models, validators, and generator into src/target_model
- update standalone target_models scripts and tests to use the shared package
- remove obsolete standalone project-side conformance code and refresh the standalone spec/docs
