+++ src/target_model/__init__.py
@@ -1,0 +1,22 @@
+"""Shared target-model domain types and validators."""
+
+from src.target_model.conformance import ConformanceIssue, TargetModelConformanceValidator
+from src.target_model.models import ColumnSpec, EntitySpec, ForeignKeySpec, GlobalConstraint, ModelMetadata, NamingConventions, TargetModel
+from src.target_model.spec_validator import SpecValidationIssue, TargetModelSpecValidator
+from src.target_model.template_generator import generate_project_template, render_project_template_yaml
+
+__all__ = [
+    "ColumnSpec",
+    "ConformanceIssue",
+    "EntitySpec",
+    "ForeignKeySpec",
+    "GlobalConstraint",
+    "ModelMetadata",
+    "NamingConventions",
+    "SpecValidationIssue",
+    "TargetModel",
+    "TargetModelConformanceValidator",
+    "TargetModelSpecValidator",
+    "generate_project_template",
+    "render_project_template_yaml",
+]
