--- target_models/src/target_model_spec/__init__.py
@@ -1,16 +1,0 @@
-from target_model_spec.models import ColumnSpec, EntitySpec, ForeignKeySpec, GlobalConstraint, ModelMetadata, NamingConventions, TargetModel
-from target_model_spec.template_generator import generate_project_template, render_project_template_yaml
-from target_model_spec.validator import TargetModelSpecValidator
-
-__all__ = [
-    "ColumnSpec",
-    "EntitySpec",
-    "ForeignKeySpec",
-    "GlobalConstraint",
-    "ModelMetadata",
-    "NamingConventions",
-    "TargetModel",
-    "TargetModelSpecValidator",
-    "generate_project_template",
-    "render_project_template_yaml",
-]
